Rename mysql package to database, add mysql example
